requirements.txt
# Python Dependencies for GoodFoods AI Reservation System

# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM API Client - Groq (Fast & Free Llama)
groq>=0.9.0          # Groq API for Llama-3.3-70B

# Data Processing
pydantic>=2.0.0        # Data validation
python-dateutil>=2.8.2 # Date/time parsing

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development Tools
black>=23.0.0          # Code formatting
flake8>=6.0.0          # Linting

# Utilities
typing-extensions>=4.8.0
